requirements.txt
openai==0.28
streamlit
python-dotenv
google-api-python-client
bs4
fpdf
